refactor(errand): 의뢰 엔티티 디렉토리 이동 domain -> domain.entity
related: #36
