+++ src/main/java/com/zerobase/nsbackend/errand/domain/repository/ErrandRepository.java
@@ -1,6 +1,6 @@
 package com.zerobase.nsbackend.errand.domain.repository;
 
-import com.zerobase.nsbackend.errand.domain.Errand;
+import com.zerobase.nsbackend.errand.domain.entity.Errand;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ErrandRepository extends JpaRepository<Errand, Long> {
