+++ src/main/java/com/zerobase/nsbackend/errand/web/ErrandController.java
@@ -1,6 +1,6 @@
 package com.zerobase.nsbackend.errand.web;
 
-import com.zerobase.nsbackend.errand.domain.Errand;
+import com.zerobase.nsbackend.errand.domain.entity.Errand;
 import com.zerobase.nsbackend.errand.domain.ErrandService;
 import com.zerobase.nsbackend.errand.dto.ErrandCreateRequest;
 import com.zerobase.nsbackend.errand.dto.ErrandDto;
